aswapp/requirements.txt
asgiref==3.2.10
certifi==2020.11.8
cffi==1.14.3
chardet==3.0.4
cryptography==3.2.1
defusedxml==0.7.0rc1
dj-database-url==0.5.0
Django==3.1.2
django-braces==1.14.0
django-heroku==0.3.1
django-oauth-toolkit==1.3.3
django-oauth2-provider==0.2.6.1
django-rest-framework-social-oauth2==1.1.0
djangorestframework==3.12.2
djangorestframework-api-key==2.0.0
idna==2.10
oauth2-provider==0.0
oauthlib==3.1.0
psycopg2==2.8.6
pycparser==2.20
PyJWT==1.7.1
python-dotenv==0.15.0
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
shortuuid==1.0.1
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
sqlparse==0.4.1
urllib3==1.25.11
waitress==1.4.4
whitenoise==5.2.0
